Save file.

+++ Mathematics/NumPy/dtype.py
@@ -1,5 +1,5 @@
 import numpy as np
-a = np.array([1, 4, 5, 8], int)
+a = np.array([1, 4, 5, 8], float)
 print a
 print type(a)
 print a[0]
